Refactored from test_int128 to test_convert_int128
