+++ tests/parser/functions/test_convert_int128.py
@@ -1,3 +1,53 @@
+from vyper.exceptions import (
+    InvalidLiteralException
+)
+
+
+def test_convert_bytes_to_int128(assert_compile_failed, get_contract_with_gas_estimation):
+    # Test valid bytes input for conversion
+    test_success = """
+@public
+def foo(bar: bytes[5]) -> int128:
+    return convert(bar, int128)
+    """
+
+    c = get_contract_with_gas_estimation(test_success)
+    assert c.foo(b'\x00\x00\x00\x00\x00') == 0
+    assert c.foo(b'\x00\x07\x5B\xCD\x15') == 123456789
+
+    test_success = """
+@public
+def foo(bar: bytes[32]) -> int128:
+    return convert(bar, int128)
+    """
+
+    c = get_contract_with_gas_estimation(test_success)
+    assert c.foo(b'\x00' * 32) == 0
+    assert c.foo(b'\xff' * 32) == -1
+
+    # Test overflow bytes input for conversion
+    test_fail = """
+@public
+def foo(bar: bytes[33]) -> int128:
+    return convert(bar, int128)
+    """
+
+    assert_compile_failed(
+        lambda: get_contract_with_gas_estimation(test_fail),
+        InvalidLiteralException
+    )
+
+    test_fail = """
+@public
+def foobar() -> int128:
+    barfoo: bytes[63] = "Hello darkness, my old friend I've come to talk with you again."
+    return convert(barfoo, int128)
+    """
+
+    assert_compile_failed(
+        lambda: get_contract_with_gas_estimation(test_fail),
+        InvalidLiteralException
+    )
 
 
 def test_convert_bytes32_to_num_overflow(assert_tx_failed, get_contract_with_gas_estimation):
